+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # 1.1.0 (2015-12-28)
 
 ## Features
+
 - add in 4 new types: facebookMessage, gmail, email, linkedIn
 - add in default svg's and styling
 - refactor to support custom HTML for buttons
 - refactor to prepare for url shortening on click (asynchronous)
+
+## Breaking Changes
+
+- remove React support (did not maintain it with the updates above and do not want to bloat this package with extra dependencies).